requirements.txt
flask>=2.0.0
numpy>=1.21.0
pillow>=9.0.0
werkzeug>=2.0.0
scikit-learn>=1.0.0
joblib>=1.0.0
gunicorn